refactor: modelディレクトリを削除してschemaファイルに型を統合
- src/domain/article/model/ ディレクトリを完全削除
- Article型とReadHistory型をschemaファイルに移行
- 全import文をmodelからschemaに変更 (14ファイル)
- index.tsのexportをdefault exportからtype exportに変更
- 型定義の重複を解消し、schemaファイルに集約
- ディレクトリ構造をより簡潔に整理
