+++ src/domain/article/index.ts
@@ -1,8 +1,7 @@
 export { default as createArticleService } from './factory/articleServiceFactory'
-export { default as Article } from './model/article'
-export { default as ReadHistory } from './model/readHistory'
 export type { ArticleQueryParams } from './schema/articleQuerySchema'
 export { articleQuerySchema } from './schema/articleQuerySchema'
+export type { Article } from './schema/articleSchema'
 export type {
   ArticleIdParam,
   CreateReadHistoryApiInput,
