+++ src/domain/article/infrastructure/articleQueryServiceImpl.ts
@@ -8,9 +8,9 @@ import {
 } from '@/common/pagination'
 import { AsyncResult, Nullable, resultError, resultSuccess } from '@/common/types/utility'
 import fromPrismaToArticle from '@/domain/article/infrastructure/articleMapper'
-import Article from '@/domain/article/model/article'
 import { ArticleQueryService } from '@/domain/article/repository/articleQueryService'
 import { ArticleQueryParams } from '@/domain/article/schema/articleQuerySchema'
+import type { Article } from '@/domain/article/schema/articleSchema'
 import { RdbClient } from '@/infrastructure/rdb'
 
 export default class ArticleQueryServiceImpl implements ArticleQueryService {
