--- src/domain/article/model/article.ts
@@ -1,5 +1,0 @@
-import { z } from 'zod'
-import { articleSchema } from '../schema/articleSchema'
-
-export type Article = z.infer<typeof articleSchema>
-export default Article
