--- src/domain/article/model/readHistory.ts
@@ -1,5 +1,0 @@
-import { z } from 'zod'
-import { readHistorySchema } from '../schema/readHistorySchema'
-
-export type ReadHistory = z.infer<typeof readHistorySchema>
-export default ReadHistory
